requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorclip==1.0.6
decorator==4.4.2
Django==5.1.1
Faker==30.1.0
fpdf==1.7.2
idna==3.10
image==1.5.33
imageio==2.35.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
moviepy==1.0.3
numpy==2.1.1
packaging==24.1
pillow==10.4.0
playwright==1.47.0
pluggy==1.5.0
proglog==0.1.10
pydub==0.25.1
pyee==12.0.0
pytest==8.3.3
pytest-rerunfailures==14.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3